# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.AddForTrackDTO:
    properties:
      genres:
        items:
          type: string
        type: array
    type: object
...
  github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.CommentDTO:
    properties:
      id:
        type: string
      stars:
        type: integer
      text:
        type: string
      track_id:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.GenreDTO:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.MusicianDTO:
    properties:
      country:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.PostCommentDTO:
    properties:
      stars:
        type: integer
      text:
        type: string
    type: object
...
  github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.RegisterMusicianDTO:
    properties:
      country:
        type: string
      description:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - country
    - description
    - email
    - name
    - password
    type: object
...
  github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.TrackDTO:
    properties:
      album_id:
        type: string
      id:
        type: string
      name:
        type: string
      url:
        type: string
    type: object
...
  internal_adapters_delivery_api.RestErrorBadRequest:
    properties:
      error:
        example: bad request
        type: string
      status:
        example: 400
        type: integer
      timestamp:
        example: "2020-11-10T23:00:00+00:00"
        type: string
    type: object
  internal_adapters_delivery_api.RestErrorConflict:
    properties:
      error:
        example: conflict
        type: string
      status:
        example: 409
        type: integer
      timestamp:
        example: "2020-11-10T23:00:00+00:00"
        type: string
    type: object
  internal_adapters_delivery_api.RestErrorForbidden:
    properties:
      error:
        example: forbidden
        type: string
      status:
        example: 403
        type: integer
      timestamp:
        example: "2020-11-10T23:00:00+00:00"
        type: string
    type: object
  internal_adapters_delivery_api.RestErrorInternalError:
    properties:
      error:
        example: internal server error
        type: string
      status:
        example: 500
        type: integer
      timestamp:
        example: "2020-11-10T23:00:00+00:00"
        type: string
    type: object
  internal_adapters_delivery_api.RestErrorNotFound:
    properties:
      error:
        example: not found
        type: string
      status:
        example: 404
        type: integer
      timestamp:
        example: "2020-11-10T23:00:00+00:00"
        type: string
    type: object
  internal_adapters_delivery_api.RestErrorUnauthorized:
    properties:
      error:
        example: unauthorized
        type: string
      status:
        example: 401
        type: integer
      timestamp:
        example: "2020-11-10T23:00:00+00:00"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a Sigma Music API.
  termsOfService: http://swagger.io/terms/
  title: Sigma Music API
  version: "1.0"
paths:
...
  /genres:
    get:
      consumes:
      - application/json
      description: get all genres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.GenreDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      summary: GetAllGenres
      tags:
      - genre
  /genres/{id}:
    get:
      consumes:
      - application/json
      description: get genre by id
      parameters:
      - description: genre id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.GenreDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      summary: GetGenreByID
      tags:
      - genre
  /musicians:
    get:
      consumes:
      - application/json
      description: get all musicians
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.MusicianDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      summary: GetAllMusicians
      tags:
      - musician
...
  /musicians/{musician_id}:
    get:
      consumes:
      - application/json
      description: get musician by id
      parameters:
      - description: musician id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.MusicianDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      summary: GetMusicianByID
      tags:
      - musician
...
  /musicians/register:
    post:
      consumes:
      - application/json
      description: registration
      parameters:
      - description: musician information
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.RegisterMusicianDTO'
      produces:
      - application/json
      responses:
        "201":
          description: message
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorConflict'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      summary: MusicianRegister
      tags:
      - musician
...
  /tracks/{id}/comments:
    get:
      consumes:
      - application/json
      description: get comments by track id
      parameters:
      - description: track id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.TrackDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      summary: GetByTrackID
      tags:
      - comment
    post:
      consumes:
      - application/json
      description: post comment
      parameters:
      - description: track id
        in: path
        name: id
        required: true
        type: string
      - description: post comment info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.PostCommentDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.CommentDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorUnauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      security:
      - ApiKeyAuth: []
      summary: PostComment
      tags:
      - comment
  /tracks/{track_id}/genres:
    get:
      consumes:
      - application/json
      description: get genres by track id
      parameters:
      - description: track id
        in: path
        name: track_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.GenreDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      summary: GetGenresByTrackID
      tags:
      - genre
    patch:
      consumes:
      - application/json
      description: set genres for track
      parameters:
      - description: add genres payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.AddForTrackDTO'
      - description: track id
        in: path
        name: track_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorUnauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorForbidden'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      security:
      - ApiKeyAuth: []
      summary: SetGenresForTrack
      tags:
      - genre
...
  /users/me/comments:
    get:
      consumes:
      - application/json
      description: get comments by user id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_hanoys_sigma-music_internal_adapters_delivery_api_dto.TrackDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapters_delivery_api.RestErrorInternalError'
      summary: GetCommentByUserID
      tags:
      - comment
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"